requirements.txt
autograd==1.8.0
blessed==1.21.0
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.1
cycler==0.12.1
filelock==3.17.0
fonttools==4.57.0
fsspec==2025.2.0
git-filter-repo==2.47.0
gpustat==1.1.1
GPUtil==1.4.0
hf-xet==1.1.5
huggingface-hub==0.33.4
idna==3.10
iperf3==0.1.11
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.12.0.88
packaging==24.2
pillow==11.1.0
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-pretrained-vit==0.0.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sympy==1.13.1
threadpoolctl==3.5.0
timm==1.0.17
tokenizers==0.21.2
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.53.2
triton==3.2.0
typing_extensions==4.12.2
urllib3==2.5.0
wcwidth==0.2.13
